+++ viewer-core/src/main/java/org/edmcouncil/spec/fibo/weasel/ontology/updater/UpdateBlocker.java
@@ -5,8 +5,25 @@ import org.springframework.stereotype.Component;
 @Component
 public class UpdateBlocker {
 
+  private Boolean initializeAppDone = false;
   private Boolean block;
   private Boolean updateNow = false;
+
+  public Boolean isInitializeAppDone() {
+    return initializeAppDone;
+  }
+
+  public void setInitializeAppDone(Boolean initializeAppDone) {
+    this.initializeAppDone = initializeAppDone;
+  }
+
+  public Boolean getBlock() {
+    return block;
+  }
+
+  public void setBlock(Boolean block) {
+    this.block = block;
+  }
 
   void setBlockerStatus(Boolean status) {
     this.block = status;