+++ viewer-core/src/main/java/org/edmcouncil/spec/fibo/weasel/ontology/updater/model/UpdateJobStatus.java
@@ -2,5 +2,5 @@ package org.edmcouncil.spec.fibo.weasel.ontology.updater.model;
 
 
 public enum UpdateJobStatus {
-WAITING, CREATED, IN_PROGRESS, DONE, ERROR
+WAITING, CREATED, IN_PROGRESS, DONE, ERROR, INTERRUPT_IN_PROGRESS
 }