+++ web-app/src/main/java/org/edmcouncil/spec/fibo/view/controller/HintController.java
@@ -6,9 +6,11 @@ import java.util.Optional;
 import org.edmcouncil.spec.fibo.view.service.TextSearchService;
 import org.edmcouncil.spec.fibo.view.util.UrlChecker;
 import org.edmcouncil.spec.fibo.weasel.ontology.searcher.model.hint.HintItem;
+import org.edmcouncil.spec.fibo.weasel.ontology.updater.UpdateBlocker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,7 +24,7 @@ import org.springframework.web.bind.annotation.RequestParam;
  * @author Michał Daniel (michal.daniel@makolab.com)
  */
 @Controller
-@RequestMapping(value = {"/hint"})
+@RequestMapping(value = {"/api/hint"})
 public class HintController {
 
   private static final Logger LOG = LoggerFactory.getLogger(HintController.class);
@@ -30,11 +32,17 @@ public class HintController {
 
   @Autowired
   private TextSearchService textSearch;
+  @Autowired
+  private UpdateBlocker blocker;
 
   @PostMapping(value = {"", "/max/{max}"})
   public ResponseEntity getHints(
       @RequestBody String query,
       @PathVariable Optional<Integer> max) {
+    if (!blocker.isInitializeAppDone()) {
+      LOG.debug("Application initialization has not completed");
+      return new ResponseEntity<>("503 Service Unavailable", HttpStatus.SERVICE_UNAVAILABLE);
+    }
     Integer maxHintCount = max.isPresent() ? max.get() : DEFAULT_MAX_HINT_RESULT_COUNT;
     LOG.debug("[REQ] POST hint | query =  {{}}  | max = {{}}", query, maxHintCount);